requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
motor
boto3
python-dotenv
